Split FromGhc into submodules
Extract 8 submodules from the 1521-line FromGhc.hs:
Internal, Doc, Exports, Names, ItemKind, Constructors,
InstanceParents, Merge. No re-exports; callers import
submodules directly. fromGhc entry point stays in FromGhc.hs.
